+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Show overdue search history when search query is empty
 - Load overdue search history when overdue search screen is opened
 - Add overdue search query validator
+- Validate search query when overdue search query is changed
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291